padding changes, content reorder and app clean
