+++ .github/workflows/gh-page.yml
@@ -1,5 +1,5 @@
 # Simple workflow for deploying static content to GitHub Pages
-name: Deploy static content to Pages
+name: Deploy PodPal website app
 
 on:
   workflow_dispatch:
